# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: LupexWallet API
  version: "1.0.0"
  description: >
    REST API однопользовательского приложения учета личных финансов.
    Технические идентификаторы — на английском, см. docs/architecture/ddd-model.md
    для доменной модели и docs/database/schema.md для схемы БД.
  license:
    name: Proprietary — внутренний проект, не публикуется
servers:
  - url: /api/v1
security:
  - BearerAuth: []
...
paths:
...
  /wallet-types:
    get:
      tags: [WalletTypes]
      summary: Список типов кошельков
      operationId: listWalletTypes
      parameters:
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: includeInactive
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReferenceItemPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [WalletTypes]
      summary: Добавить тип кошелька (UC-28, раздел 7)
      operationId: createWalletType
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReferenceItemCreateRequest" }
      responses:
        "201":
          description: Создан
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReferenceItem" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /wallet-types/{id}/deactivate:
    parameters: [{ $ref: "#/components/parameters/IdParam" }]
    post:
      tags: [WalletTypes]
      summary: Деактивировать тип кошелька (решено, Q11)
      operationId: deactivateWalletType
      responses:
        "200":
          description: Деактивирован
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReferenceItem" }
        "404": { $ref: "#/components/responses/NotFound" }
  /wallet-types/{id}:
    parameters: [{ $ref: "#/components/parameters/IdParam" }]
    delete:
      tags: [WalletTypes]
      summary: Удалить неиспользуемый тип кошелька
      operationId: deleteWalletType
      responses:
        "204": { description: Удален }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Тип уже используется хотя бы одним кошельком — используйте деактивацию.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /operation-types:
    get:
      tags: [OperationTypes]
      summary: Список типов операций
      operationId: listOperationTypes
      parameters:
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: includeInactive
          in: query
          schema: { type: boolean, default: false }
        - name: behaviorKindId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperationTypePage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [OperationTypes]
      summary: Добавить тип операции (UC-15)
      operationId: createOperationType
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OperationTypeCreateRequest" }
      responses:
        "201":
          description: Создан
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperationType" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /operation-types/{id}/deactivate:
    parameters: [{ $ref: "#/components/parameters/IdParam" }]
    post:
      tags: [OperationTypes]
      summary: Деактивировать тип операции
      operationId: deactivateOperationType
      responses:
        "200":
          description: Деактивирован
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperationType" }
        "404": { $ref: "#/components/responses/NotFound" }
  /operation-types/{id}:
    parameters: [{ $ref: "#/components/parameters/IdParam" }]
    delete:
      tags: [OperationTypes]
      summary: Удалить неиспользуемый тип операции
      operationId: deleteOperationType
      responses:
        "204": { description: Удален }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Тип уже используется хотя бы одной операцией — используйте деактивацию.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /operation-behavior-kinds:
    get:
      tags: [OperationBehaviorKinds]
      summary: Системный справочник видов поведения операции (только чтение)
      operationId: listOperationBehaviorKinds
      description: >
        Income | Expense | Transfer | Adjustment. Без create/update/delete —
        набор фиксирован при развертывании (ddd-model.md, §2.3).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OperationBehaviorKind" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /currencies:
    get:
      tags: [Currencies]
      summary: Список валют
      operationId: listCurrencies
      parameters:
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: includeInactive
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrencyPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Currencies]
      summary: Добавить валюту (в т.ч. криптовалюту, решено при проектировании домена)
      operationId: createCurrency
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CurrencyCreateRequest" }
      responses:
        "201":
          description: Создана
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Currency" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: Код валюты уже занят.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /currencies/{id}/deactivate:
    parameters: [{ $ref: "#/components/parameters/IdParam" }]
    post:
      tags: [Currencies]
      summary: Деактивировать валюту
      operationId: deactivateCurrency
      responses:
        "200":
          description: Деактивирована
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Currency" }
        "404": { $ref: "#/components/responses/NotFound" }
  /currencies/{id}:
    parameters: [{ $ref: "#/components/parameters/IdParam" }]
    delete:
      tags: [Currencies]
      summary: Удалить неиспользуемую валюту
      operationId: deleteCurrency
      responses:
        "204": { description: Удалена }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Валюта уже используется — используйте деактивацию.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

  # ---------- Operations ----------
...
  /reporting/total-amount:
    get:
      tags: [Reporting]
      summary: Общая сумма — текущая или на дату (UC-19, UC-21)
      operationId: getTotalAmount
      parameters:
        - name: date
          in: query
          description: Без параметра — текущая сумма (UC-19); с параметром — историческая (UC-21).
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TotalAmount" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Нет основного кошелька (общую сумму считать не к чему).
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
      # 404 здесь — штатный случай (нет основного кошелька), не только "запись не найдена по
      # id"; схема ответа та же, что у components/responses/NotFound (application/problem+json,
      # Problem), описание переопределено под конкретный смысл этого эндпоинта.

  # ---------- Audit ----------
...
components:
  parameters:
...
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    CursorParam:
      name: cursor
      in: query
      description: Непрозрачный курсор для постраничной навигации.
      schema: { type: string }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100, minimum: 1 }
  schemas:
...
    Money:
      type: object
      required: [amount, currencyId]
      properties:
        amount:
          type: string
          format: decimal
          description: Произвольная точность (ADR-0005); передается строкой, чтобы не терять точность в JSON-числах.
          example: "1234.56789012"
        currencyId: { type: string, format: uuid }
        currencyCode: { type: string, readOnly: true, example: "USD" }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    CursorPageMeta:
      type: object
      required: [nextCursor, hasMore]
      properties:
        nextCursor: { type: [string, "null"] }
        hasMore: { type: boolean }

    # ---- Auth ----
...
    ReferenceItem:
      type: object
      required: [id, name, isActive]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        name: { type: string, maxLength: 200 }
        isActive: { type: boolean, readOnly: true }
    ReferenceItemCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 200 }
    ReferenceItemPage:
      type: object
      required: [data, pagination]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/ReferenceItem" } }
        pagination: { $ref: "#/components/schemas/CursorPageMeta" }
    OperationBehaviorKind:
      type: object
      required: [id, code, name]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        code:
          type: string
          enum: [Income, Expense, Transfer, Adjustment]
        name: { type: string }
    OperationType:
      type: object
      required: [id, name, behaviorKindId, isActive]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        name: { type: string, maxLength: 200 }
        behaviorKindId: { type: string, format: uuid }
        isActive: { type: boolean, readOnly: true }
    OperationTypeCreateRequest:
      type: object
      required: [name, behaviorKindId]
      properties:
        name: { type: string, maxLength: 200 }
        behaviorKindId: { type: string, format: uuid }
    OperationTypePage:
      type: object
      required: [data, pagination]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/OperationType" } }
        pagination: { $ref: "#/components/schemas/CursorPageMeta" }
    Currency:
      type: object
      required: [id, code, name, isActive]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        code: { type: string, maxLength: 16, example: "BTC" }
        name: { type: string, maxLength: 200, example: "Bitcoin" }
        isActive: { type: boolean, readOnly: true }
    CurrencyCreateRequest:
      type: object
      required: [code, name]
      properties:
        code: { type: string, maxLength: 16 }
        name: { type: string, maxLength: 200 }
    CurrencyPage:
      type: object
      required: [data, pagination]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Currency" } }
        pagination: { $ref: "#/components/schemas/CursorPageMeta" }

    # ---- Operations ----
...
    TotalAmount:
      type: object
      required: [amount, ratesAsOfDate, excludedWallets]
      properties:
        date:
          type: [string, "null"]
          format: date
          description: Запрошенная дата; null означает «текущая сумма» (UC-19).
        amount: { $ref: "#/components/schemas/Money" }
        ratesAsOfDate:
          type: string
          format: date
          description: Дата курса, реально примененного при конвертации (может отличаться от запрошенной, если точного курса нет — раздел 5).
        excludedWallets:
          type: array
          description: >-
            Кошельки, исключенные из amount из-за отсутствия курса для их валюты
            (валюта не поддерживается/не получена источником курсов) — решение по риску №6
            (docs/PROGRESS.md, П.5.1): исключение из суммы с явной пометкой является основным
            путем, не запасным; не 409, не курс 1:1.
          items: { $ref: "#/components/schemas/ExcludedWallet" }
    ExcludedWallet:
      type: object
      required: [walletId, reason]
      properties:
        walletId: { type: string, format: uuid }
        reason: { type: string, description: "Человекочитаемая причина исключения кошелька из суммы." }

    # ---- Audit ----
...
  responses:
    BadRequest:
      description: Некорректные параметры запроса
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Отсутствует или неверен токен доступа
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Запись не найдена
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT